Refactored ConversationConditions and fixed state issues
